requirements.txt
# AWD Policy Transportability Analysis - Python Dependencies
# Installation: pip install -r requirements.txt

# Core Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Geospatial Processing
rasterio==1.3.6
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0

# Google Earth Engine
earthengine-api==0.1.380

# Visualization & Plotting
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0

# Configuration Management
PyYAML==6.0

# Scientific Computing
scikit-learn==1.3.0
scikit-image==0.21.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Development & Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.4.1
